requirements.txt
# Core dependencies for Autonomous SLAM Robot
# Install with: pip install -r requirements.txt

# LiDAR
rplidar-roboticia>=1.0.0

# Motor Control
gpiozero>=2.0.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.6.0

# Serial Communication (usually included with rplidar)
pyserial>=3.5

